+++ config/releases.exs
@@ -25,6 +25,9 @@ config :dealog_backoffice, DealogBackoffice.EventStore,
   url: event_store_database_url,
   pool_size: String.to_integer(System.get_env("POOL_SIZE") || "10")
 
+config :dealog_backoffice, :projection,
+  projectors: String.to_atom(System.get_env("PROJECTORS")) || :all
+
 secret_key_base =
   System.get_env("SECRET_KEY_BASE") ||
     raise """